requirements.txt
asgiref==3.8.1
Django==5.0.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
packaging==25.0
pillow==10.2.0
psycopg2-binary==2.9.9
PyJWT==2.10.1
python-dotenv==1.0.1
pytz==2025.2
sqlparse==0.5.3
whitenoise==6.9.0
boto3==1.38.37
django-storages==1.14.6
django-jazzmin==3.0.1
reportlab==4.0.4
 